split stream.h into sync_stream.h and async_stream.h
