+++ BUILD
@@ -735,6 +735,7 @@ cc_library(
     "include/grpc++/server_builder.h",
     "include/grpc++/server_context.h",
     "include/grpc++/server_credentials.h",
+    "include/grpc++/support/async_stream.h",
     "include/grpc++/support/async_unary_call.h",
     "include/grpc++/support/auth_context.h",
     "include/grpc++/support/byte_buffer.h",
@@ -746,8 +747,8 @@ cc_library(
     "include/grpc++/support/slice.h",
     "include/grpc++/support/status.h",
     "include/grpc++/support/status_code_enum.h",
-    "include/grpc++/support/stream.h",
     "include/grpc++/support/stub_options.h",
+    "include/grpc++/support/sync_stream.h",
     "include/grpc++/support/thread_pool_interface.h",
     "include/grpc++/support/time.h",
   ],
@@ -821,6 +822,7 @@ cc_library(
     "include/grpc++/server_builder.h",
     "include/grpc++/server_context.h",
     "include/grpc++/server_credentials.h",
+    "include/grpc++/support/async_stream.h",
     "include/grpc++/support/async_unary_call.h",
     "include/grpc++/support/auth_context.h",
     "include/grpc++/support/byte_buffer.h",
@@ -832,8 +834,8 @@ cc_library(
     "include/grpc++/support/slice.h",
     "include/grpc++/support/status.h",
     "include/grpc++/support/status_code_enum.h",
-    "include/grpc++/support/stream.h",
     "include/grpc++/support/stub_options.h",
+    "include/grpc++/support/sync_stream.h",
     "include/grpc++/support/thread_pool_interface.h",
     "include/grpc++/support/time.h",
   ],
