+++ Makefile
@@ -4654,6 +4654,7 @@ PUBLIC_HEADERS_CXX += \
     include/grpc++/server_builder.h \
     include/grpc++/server_context.h \
     include/grpc++/server_credentials.h \
+    include/grpc++/support/async_stream.h \
     include/grpc++/support/async_unary_call.h \
     include/grpc++/support/auth_context.h \
     include/grpc++/support/byte_buffer.h \
@@ -4665,8 +4666,8 @@ PUBLIC_HEADERS_CXX += \
     include/grpc++/support/slice.h \
     include/grpc++/support/status.h \
     include/grpc++/support/status_code_enum.h \
-    include/grpc++/support/stream.h \
     include/grpc++/support/stub_options.h \
+    include/grpc++/support/sync_stream.h \
     include/grpc++/support/thread_pool_interface.h \
     include/grpc++/support/time.h \
 
@@ -4896,6 +4897,7 @@ PUBLIC_HEADERS_CXX += \
     include/grpc++/server_builder.h \
     include/grpc++/server_context.h \
     include/grpc++/server_credentials.h \
+    include/grpc++/support/async_stream.h \
     include/grpc++/support/async_unary_call.h \
     include/grpc++/support/auth_context.h \
     include/grpc++/support/byte_buffer.h \
@@ -4907,8 +4909,8 @@ PUBLIC_HEADERS_CXX += \
     include/grpc++/support/slice.h \
     include/grpc++/support/status.h \
     include/grpc++/support/status_code_enum.h \
-    include/grpc++/support/stream.h \
     include/grpc++/support/stub_options.h \
+    include/grpc++/support/sync_stream.h \
     include/grpc++/support/thread_pool_interface.h \
     include/grpc++/support/time.h \
