+++ include/grpc++/generic/async_generic_service.h
@@ -35,7 +35,7 @@
 #define GRPCXX_GENERIC_ASYNC_GENERIC_SERVICE_H
 
 #include <grpc++/support/byte_buffer.h>
-#include <grpc++/support/stream.h>
+#include <grpc++/support/async_stream.h>
 
 struct grpc_server;
