+++ tools/run_tests/sources_and_headers.json
@@ -13143,6 +13143,7 @@
       "include/grpc++/server_builder.h", 
       "include/grpc++/server_context.h", 
       "include/grpc++/server_credentials.h", 
+      "include/grpc++/support/async_stream.h", 
       "include/grpc++/support/async_unary_call.h", 
       "include/grpc++/support/auth_context.h", 
       "include/grpc++/support/byte_buffer.h", 
@@ -13154,8 +13155,8 @@
       "include/grpc++/support/slice.h", 
       "include/grpc++/support/status.h", 
       "include/grpc++/support/status_code_enum.h", 
-      "include/grpc++/support/stream.h", 
       "include/grpc++/support/stub_options.h", 
+      "include/grpc++/support/sync_stream.h", 
       "include/grpc++/support/thread_pool_interface.h", 
       "include/grpc++/support/time.h", 
       "src/cpp/client/create_channel_internal.h", 
@@ -13192,6 +13193,7 @@
       "include/grpc++/server_builder.h", 
       "include/grpc++/server_context.h", 
       "include/grpc++/server_credentials.h", 
+      "include/grpc++/support/async_stream.h", 
       "include/grpc++/support/async_unary_call.h", 
       "include/grpc++/support/auth_context.h", 
       "include/grpc++/support/byte_buffer.h", 
@@ -13203,8 +13205,8 @@
       "include/grpc++/support/slice.h", 
       "include/grpc++/support/status.h", 
       "include/grpc++/support/status_code_enum.h", 
-      "include/grpc++/support/stream.h", 
       "include/grpc++/support/stub_options.h", 
+      "include/grpc++/support/sync_stream.h", 
       "include/grpc++/support/thread_pool_interface.h", 
       "include/grpc++/support/time.h", 
       "src/cpp/client/channel.cc", 
@@ -13315,6 +13317,7 @@
       "include/grpc++/server_builder.h", 
       "include/grpc++/server_context.h", 
       "include/grpc++/server_credentials.h", 
+      "include/grpc++/support/async_stream.h", 
       "include/grpc++/support/async_unary_call.h", 
       "include/grpc++/support/auth_context.h", 
       "include/grpc++/support/byte_buffer.h", 
@@ -13326,8 +13329,8 @@
       "include/grpc++/support/slice.h", 
       "include/grpc++/support/status.h", 
       "include/grpc++/support/status_code_enum.h", 
-      "include/grpc++/support/stream.h", 
       "include/grpc++/support/stub_options.h", 
+      "include/grpc++/support/sync_stream.h", 
       "include/grpc++/support/thread_pool_interface.h", 
       "include/grpc++/support/time.h", 
       "src/cpp/client/create_channel_internal.h", 
@@ -13361,6 +13364,7 @@
       "include/grpc++/server_builder.h", 
       "include/grpc++/server_context.h", 
       "include/grpc++/server_credentials.h", 
+      "include/grpc++/support/async_stream.h", 
       "include/grpc++/support/async_unary_call.h", 
       "include/grpc++/support/auth_context.h", 
       "include/grpc++/support/byte_buffer.h", 
@@ -13372,8 +13376,8 @@
       "include/grpc++/support/slice.h", 
       "include/grpc++/support/status.h", 
       "include/grpc++/support/status_code_enum.h", 
-      "include/grpc++/support/stream.h", 
       "include/grpc++/support/stub_options.h", 
+      "include/grpc++/support/sync_stream.h", 
       "include/grpc++/support/thread_pool_interface.h", 
       "include/grpc++/support/time.h", 
       "src/cpp/client/channel.cc",
